Cleaned up code and added three more components.
Added about me component to display fun facts about me.
Added a Menu component to display a menu for the user to click on.
Added some styling to the menu component to allow the user to switch between pages by clicking on the options.
Removed files from the database that were a remnant from my other project.
Started creating the schema for my database.